requirements.txt
aiogram==3.12.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
ujson==5.10.0
fastapi==0.115.0
uvicorn==0.32.0
# Database
asyncpg==0.29.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9  # Sync driver for alembic migrations